Add chainCluster DSL to testclusters
With this change tasks get a `chainCluster` method that accepts a Task.
This is a short-hand to express that the task depends on the other and
needs to use all it's cluers and it's dependencies clusters.

As a show-case this is applied to CCR tests, but it will also be usefull
in BWC tests to eliminate boilerplate.

+++ buildSrc/src/main/java/org/elasticsearch/gradle/testclusters/ElasticsearchNode.java
@@ -365,8 +365,12 @@ public class ElasticsearchNode implements TestClusterConfiguration {
         LOGGER.info("Restarting {}", this);
         stop(false);
         try {
-            Files.delete(httpPortsFile);
-            Files.delete(transportPortFile);
+            if (Files.exists(httpPortsFile)) {
+                Files.delete(httpPortsFile);
+            }
+            if (Files.exists(transportPortFile)) {
+                Files.delete(transportPortFile);
+            }
         } catch (IOException e) {
             throw new UncheckedIOException(e);
         }
